scripts/requirements-test.txt
groq==0.4.1
python-dotenv==1.0.0
